I think collsion bouncing is working but unsure if it is working yet. Need to hook it up to game objects so can visually see objects colliding.

+++ include/PhysicsEngine.h
@@ -16,7 +16,7 @@ public:
 
 	void addObject(PhysicsObject* object);
 	
-	void simulate(long delta);
+	void simulate(double delta);
 
 	void handleCollisions();
